feat : 재생 & 일시정지 기능, seek bar 로 재생 지점 선택 기능 추가

+++ MusicPlayerSample-Programmers/Domain/Entities/Music.swift
@@ -16,5 +16,5 @@ struct Music {
     let albumImage: UIImage
     let file: Data
     let lyrics: Lyrics
-    let duration: Int
+    let duration: Float
 }
